Modified to match the new battle mechanic

+++ src2/OutsideEntities/Monsters/Monster.java
@@ -20,13 +20,13 @@ public class Monster extends AbstractCharacter {
     private List<State> states;
     private int damage;
     private String message;
-    private float physical_def;
-    private float magical_def;
+    private double physical_def;
+    private double magical_def;
     private boolean flying;
     private int gold_drop;
 
 
-    public Monster(String name, int health, int damage, float physical_def, float magical_def, boolean flying, int gold_drop) {
+    public Monster(String name, int health, int damage, double physical_def, double magical_def, boolean flying, int gold_drop) {
         super(name, health);
         this.states = new ArrayList<State>();
         this.damage = damage;
@@ -39,9 +39,9 @@ public class Monster extends AbstractCharacter {
 
     public int getDamage() {return this.damage;}
 
-    public float getPhysicalDef() {return this.physical_def;}
+    public double getPhysicalDef() {return this.physical_def;}
 
-    public float getMagicalDef() {return this.magical_def;}
+    public double getMagicalDef() {return this.magical_def;}
 
     public boolean getFlying() {return this.flying;}
